Split World into BlueprintWorld and World
