+++ blade-engine/src/com/bladecoder/engine/model/BlueprintWorld.java
@@ -1,0 +1,49 @@
+package com.bladecoder.engine.model;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+import java.util.Collection;
+
+public class BlueprintWorld {
+	@JsonProperty
+	protected int width;
+	@JsonProperty
+	protected int height;
+	@JsonProperty
+	protected String initChapter;
+
+	@JsonProperty
+	private Collection<Verb> getVerbs() {
+		return VerbManager.getWorldVerbs().values();
+	}
+
+	private void setVerbs(Collection<Verb> verbs) {
+		for (Verb verb : verbs) {
+			VerbManager.addDefaultVerb(verb.getId(), verb);
+		}
+	}
+
+	public int getWidth() {
+		return width;
+	}
+
+	public void setWidth(int width) {
+		this.width = width;
+	}
+
+	public int getHeight() {
+		return height;
+	}
+
+	public void setHeight(int height) {
+		this.height = height;
+	}
+
+	public String getInitChapter() {
+		return initChapter;
+	}
+
+	public void setInitChapter(String initChapter) {
+		this.initChapter = initChapter;
+	}
+}
